정은화 - Book Controller - memeber 세션 검사 추가

+++ test26/src/main/java/ksmart/project/test26/BookController.java
@@ -2,6 +2,8 @@ package ksmart.project.test26;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,7 +21,10 @@ public class BookController {
 	
 	// 목록조회
 	@RequestMapping(value="/book/bookList", method=RequestMethod.GET)
-	public String bookList(Model model) {
+	public String bookList(Model model, HttpSession session) {
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/log/login";
+		}
 		List<Book> list = bookDao.selectBookList();
 		model.addAttribute("list", list);
 		return "book/bookList";
@@ -27,20 +32,32 @@ public class BookController {
 	
 	// 입력페이지 요청
 	@RequestMapping(value="/book/bookInsert", method=RequestMethod.GET)
-	public String bookInsert() {
+	public String bookInsert(HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		return "book/bookInsert";
 	}
 	
 	// 입력 처리요청
 	@RequestMapping(value="/book/bookInsert", method=RequestMethod.POST)
-	public String bookInsert(Book book) {
+	public String bookInsert(Book book, HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		bookDao.insertBook(book);
 		return "redirect:/book/bookList";
 	}
 	
 	// 업데이트 정보요청
 	@RequestMapping(value="/book/bookUpdate", method=RequestMethod.GET)
-	public String bookUpdate(Model model, @RequestParam(value="bookId", required=true) int bookId) {
+	public String bookUpdate(Model model, HttpSession session, @RequestParam(value="bookId", required=true) int bookId) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		Book book = bookDao.updateGetBook(bookId);
 		model.addAttribute("book", book);
 		return "book/bookUpdate";
@@ -48,14 +65,22 @@ public class BookController {
 	
 	// 업데이트 처리요청
 	@RequestMapping(value="/book/bookUpdate", method=RequestMethod.POST)
-	public String bookUpdate(Book book) {
+	public String bookUpdate(Book book, HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		bookDao.updateBook(book);		
 		return "redirect:/book/bookList";
 	}
 
 	// 삭제
 	@RequestMapping(value="/book/bookDelete", method=RequestMethod.GET)
-	public String bookDelete(@RequestParam(value="bookId", required=true) int bookId) {
+	public String bookDelete(HttpSession session, @RequestParam(value="bookId", required=true) int bookId) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		bookDao.deleteBook(bookId);
 		return "redirect:/book/bookList";
 	}
