+++ test26/src/main/java/ksmart/project/test26/CityController.java
@@ -2,6 +2,8 @@ package ksmart.project.test26;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,7 +21,11 @@ public class CityController {
 	
 	// 목록조회
 	@RequestMapping(value="/city/cityList", method=RequestMethod.GET)
-	public String cityList(Model model) {
+	public String cityList(Model model, HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		List<City> list = cityDao.selectCityList();
 		model.addAttribute("list", list);
 		return "city/cityList";
@@ -27,20 +33,32 @@ public class CityController {
 	
 	// 입력페이지 요청
 	@RequestMapping(value="/city/cityInsert", method=RequestMethod.GET)
-	public String cityInsert() {
+	public String cityInsert(HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		return "city/cityInsert";
 	}
 	
 	// 입력 처리요청
 	@RequestMapping(value="/city/cityInsert", method=RequestMethod.POST)
-	public String cityInsert(City city) {
+	public String cityInsert(City city, HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		cityDao.insertCity(city);
 		return "redirect:/city/cityList";
 	}
 	
 	// 업데이트 정보요청
 	@RequestMapping(value="/city/cityUpdate", method=RequestMethod.GET)
-	public String cityUpdate(Model model, @RequestParam(value="cityId", required=true) int cityId) {
+	public String cityUpdate(Model model, HttpSession session, @RequestParam(value="cityId", required=true) int cityId) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		City city = cityDao.updateGetCity(cityId);
 		model.addAttribute("city", city);
 		return "city/cityUpdate";
@@ -48,14 +66,22 @@ public class CityController {
 	
 	// 업데이트 처리요청
 	@RequestMapping(value="/city/cityUpdate", method=RequestMethod.POST)
-	public String cityUpdate(City city) {
+	public String cityUpdate(City city, HttpSession session) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		cityDao.updateCity(city);
 		return "redirect:/city/cityList";
 	}
 	
 	// 삭제
 	@RequestMapping(value="/city/cityDelete", method=RequestMethod.GET)
-	public String cityDelete(@RequestParam(value="cityId", required=true) int cityId) {
+	public String cityDelete(HttpSession session, @RequestParam(value="cityId", required=true) int cityId) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		cityDao.deleteCity(cityId);
 		return "redirect:/city/cityList";
 	}
